% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fixed.R
\name{get_fixed}
\alias{get_fixed}
\title{This function subsets the sample-by-variable dataframe
to the "fixed" (i.e., non-permuted) data.
This subset should include ID variables, sex, and genotypic covariates.}
\usage{
get_fixed(df, id.var = id.var, sex.var = sex.var, geno.vars = geno.vars)
}
\arguments{
\item{df}{Data frame}

\item{id.var, sex.var}{String}

\item{geno.vars}{Character vector}
}
\value{
Data frame
}
\description{
This function subsets the sample-by-variable dataframe
to the "fixed" (i.e., non-permuted) data.
This subset should include ID variables, sex, and genotypic covariates.
}
