% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pairPerm.R
\name{get_pairPerm}
\alias{get_pairPerm}
\title{This function samples individuals by pairs for permuting phenotype
vectors in the paired-permutation function (sipPair.R)}
\usage{
get_pairPerm(deglst, rid.vars = rid.vars, ibd.var = ibd.var, seed = seed)
}
\arguments{
\item{deglst}{List of data frames}

\item{rid.vars}{Character vector}

\item{ibd.var}{String}

\item{seed}{Number}
}
\value{
Data frame
}
\description{
This function samples individuals by pairs for permuting phenotype
vectors in the paired-permutation function (sipPair.R)
}
