% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pairPermDF.R
\name{get_pairPermDF}
\alias{get_pairPermDF}
\title{This function recombines the fixed data and permuted data
in the paired-permutation function (sipPair.R) into a permuted
sample-by-variable data frame.}
\usage{
get_pairPermDF(fix.df, perm.df, perm.map, id.var = id.var)
}
\arguments{
\item{fix.df, perm.df, perm.map}{Data frames}

\item{id.var}{String}
}
\value{
Data frame
}
\description{
This function recombines the fixed data and permuted data
in the paired-permutation function (sipPair.R) into a permuted
sample-by-variable data frame.
}
