% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_permDF.R
\name{get_permDF}
\alias{get_permDF}
\title{This function recombines the fixed data and permuted data
into a permuted sample-by-variable data frame.}
\usage{
get_permDF(fix.df, perm.df, perm.idx)
}
\arguments{
\item{fix.df, perm.df}{Data frames}

\item{perm.idx}{Numeric vector}
}
\value{
Data frame
}
\description{
This function recombines the fixed data and permuted data
into a permuted sample-by-variable data frame.
}
