% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_permIdx.R
\name{get_permIdx}
\alias{get_permIdx}
\title{This function samples indices for permuting phenotype vectors in
the basic permutation function (sip.R) and for permuting leftover
unpaired individuals in the paired-permutation function (sipPair.R).}
\usage{
get_permIdx(df, seed = seed)
}
\arguments{
\item{df}{Data frame}

\item{seed}{Number}
}
\value{
Numeric vector
}
\description{
This function samples indices for permuting phenotype vectors in
the basic permutation function (sip.R) and for permuting leftover
unpaired individuals in the paired-permutation function (sipPair.R).
}
