% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_permute.R
\name{get_permute}
\alias{get_permute}
\title{This function subsets the sample-by-variable data frame
to the permutable data.
This subset should include phenotypes and phenotypic covariates.}
\usage{
get_permute(df, id.var = id.var, pheno.vars = pheno.vars)
}
\arguments{
\item{df}{Data frame}

\item{id.var}{String}

\item{pheno.vars}{Character vector}
}
\value{
Data frame
}
\description{
This function subsets the sample-by-variable data frame
to the permutable data.
This subset should include phenotypes and phenotypic covariates.
}
