% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relCat.R
\name{get_relCat}
\alias{get_relCat}
\title{This function infers relatedness categories in the paired-permutation
function (sip_pair.R) prior to permutation.}
\usage{
get_relCat(pairs, rid.vars = rid.vars, ibd.var = ibd.var)
}
\arguments{
\item{pairs}{Data frame}

\item{rid.vars}{Character vector}

\item{ibd.var}{String}
}
\value{
List of data frames
}
\description{
This function infers relatedness categories in the paired-permutation
function (sip_pair.R) prior to permutation.
}
