% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relPairs.R
\name{get_relPairs}
\alias{get_relPairs}
\title{This function pairs individuals in the paired-permutation function
(sip_pair.R) into the most highly-related pairs possible.}
\usage{
get_relPairs(
  df,
  id.var = id.var,
  rel.df = rel.df,
  rid.vars = rid.vars,
  ibd.var = ibd.var
)
}
\arguments{
\item{df, rel.df}{Data frames}

\item{id.var, ibd.var}{Strings}

\item{rid.vars}{Character vector}
}
\value{
Data frame
}
\description{
This function pairs individuals in the paired-permutation function
(sip_pair.R) into the most highly-related pairs possible.
}
