% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sexDF.R
\name{get_sexDF}
\alias{get_sexDF}
\title{This function divides the sample-by-variable data frame into males
and females prior to permutation within sexes.}
\usage{
get_sexDF(df = df, sex.var = sex.var, sex.val = sex.val)
}
\arguments{
\item{df}{Data frame}

\item{sex.var}{String}

\item{sex.val}{String or integer}
}
\value{
Data frame
}
\description{
This function divides the sample-by-variable data frame into males
and females prior to permutation within sexes.
}
