% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_singSamp.R
\name{get_singSamp}
\alias{get_singSamp}
\title{This function checks for uneven numbers of individuals within relatedness
category and creates a data set for permuting single samples that will not
be pair permuted (sip_pair.R).}
\usage{
get_singSamp(df, deglst, id.var = id.var)
}
\arguments{
\item{df}{Data frame}

\item{deglst}{List of data frames}

\item{id.var}{String}
}
\value{
Data frame
}
\description{
This function checks for uneven numbers of individuals within relatedness
category and creates a data set for permuting single samples that will not
be pair permuted (sip_pair.R).
}
