% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_permDF.R
\name{order_permDF}
\alias{order_permDF}
\title{This function reorders the permuted sample-by-variable data frame
to match the ID order of the primary sample-by-variable data frame.}
\usage{
order_permDF(df = df, perm = perm, id.var = id.var)
}
\arguments{
\item{df, perm}{Data frames}

\item{id.var}{String}
}
\value{
Data frame
}
\description{
This function reorders the permuted sample-by-variable data frame
to match the ID order of the primary sample-by-variable data frame.
}
