% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenotype_correlations.R
\name{plot_phenotype_correlations}
\alias{plot_phenotype_correlations}
\title{Phenotype correlation plot}
\usage{
plot_phenotype_correlations(df = NULL, pheno.vars = NULL)
}
\arguments{
\item{df}{Data frame}

\item{pheno.vars}{Character vector}
}
\value{
Plot
}
\description{
This function returns a plot showing correlation
between phenotypes. (N.B., please omit missing values
before running this function.)
}
\details{
The function requires the following inputs:

1) A sample-by-variable data frame with phenotypes.
Column names should include phenotype names.

2) A vector of phenotype names.
(e.g., pheno.vars=c("PHENO1","PHENO2",...)).
}
\examples{
plot_phenotype_correlations(df = sip_exampleData,
pheno.vars = paste0("PHENO",1:500))
}
