% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{SIP-package}
\alias{SIP-package}
\title{SIP-package internal setup}
\description{
Declares global variables and imports common functions
used throughout the SIP package. This file is executed
when the package is loaded.
}
\keyword{internal}
