% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sip_exampleData}
\alias{sip_exampleData}
\title{Single-iteration permutation phenotype and covariate data}
\format{
A data frame with 10,000 rows and 511 variables:
\describe{
  \item{FID}{Family ID.}
  \item{IID}{Individual ID.}
  \item{ANCESTRY}{Sample ancestry.}
  \item{SEX}{Sample biological sex.}
  \item{AGE}{Sample age.}
  \item{BATCH}{Genotyping batch.}
  \item{STUDY}{Study contributing sample data.}
  \item{PCs}{4 genetic principal components.}
  \item{PHENOs}{500 simulated phenotypes.}
}
}
\source{
<https://doi.org/10.21203/rs.3.rs-873449/v1>
}
\usage{
sip_exampleData
}
\description{
A sample-by-variable data frame with IDs,
covariates, and phenotypes as column names.
}
\keyword{datasets}
