% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sipPair_exampleData}
\alias{sipPair_exampleData}
\title{Single-iteration paired permutation phenotype and covariate data}
\format{
A data frame with 10,000 rows and 309 variables:
\describe{
  \item{FID}{Family ID.}
  \item{IID}{Individual ID.}
  \item{BATCH}{Genotyping batch.}
  \item{SEX}{Sample biological sex.}
  \item{AGE}{Sample age.}
  \item{PC1}{Genetic principal component 1.}
  \item{PC2}{Genetic principal component 2.}
  \item{PC3}{Genetic principal component 3.}
  \item{PC4}{Genetic principal component 4.}
  \item{PHENO1}{Simulated phenotype 1.}
  \item{PHENO2}{Simulated phenotype 2.}
  \item{PHENO3}{Simulated phenotype 3.}
  \item{PHENO4}{Simulated phenotype 4.}
  \item{PHENO5}{Simulated phenotype 5.}
  \item{PHENO6}{Simulated phenotype 6.}
  \item{PHENO7}{Simulated phenotype 7.}
  \item{PHENO8}{Simulated phenotype 8.}
  \item{PHENO9}{Simulated phenotype 9.}
  \item{PHENO10}{Simulated phenotype 10.}
  \item{PHENO11}{Simulated phenotype 11.}
  \item{PHENO12}{Simulated phenotype 12.}
  \item{PHENO13}{Simulated phenotype 13.}
  \item{PHENO14}{Simulated phenotype 14.}
  \item{PHENO15}{Simulated phenotype 15.}
  \item{PHENO16}{Simulated phenotype 16.}
  \item{PHENO17}{Simulated phenotype 17.}
  \item{PHENO18}{Simulated phenotype 18.}
  \item{PHENO19}{Simulated phenotype 19.}
  \item{PHENO20}{Simulated phenotype 20.}
  \item{PHENO21}{Simulated phenotype 21.}
  \item{PHENO22}{Simulated phenotype 22.}
  \item{PHENO23}{Simulated phenotype 23.}
  \item{PHENO24}{Simulated phenotype 24.}
  \item{PHENO25}{Simulated phenotype 25.}
  \item{PHENO26}{Simulated phenotype 26.}
  \item{PHENO27}{Simulated phenotype 27.}
  \item{PHENO28}{Simulated phenotype 28.}
  \item{PHENO29}{Simulated phenotype 29.}
  \item{PHENO30}{Simulated phenotype 30.}
  \item{PHENO31}{Simulated phenotype 31.}
  \item{PHENO32}{Simulated phenotype 32.}
  \item{PHENO33}{Simulated phenotype 33.}
  \item{PHENO34}{Simulated phenotype 34.}
  \item{PHENO35}{Simulated phenotype 35.}
  \item{PHENO36}{Simulated phenotype 36.}
  \item{PHENO37}{Simulated phenotype 37.}
  \item{PHENO38}{Simulated phenotype 38.}
  \item{PHENO39}{Simulated phenotype 39.}
  \item{PHENO40}{Simulated phenotype 40.}
  \item{PHENO41}{Simulated phenotype 41.}
  \item{PHENO42}{Simulated phenotype 42.}
  \item{PHENO43}{Simulated phenotype 43.}
  \item{PHENO44}{Simulated phenotype 44.}
  \item{PHENO45}{Simulated phenotype 45.}
  \item{PHENO46}{Simulated phenotype 46.}
  \item{PHENO47}{Simulated phenotype 47.}
  \item{PHENO48}{Simulated phenotype 48.}
  \item{PHENO49}{Simulated phenotype 49.}
  \item{PHENO50}{Simulated phenotype 50.}
  \item{PHENO51}{Simulated phenotype 51.}
  \item{PHENO52}{Simulated phenotype 52.}
  \item{PHENO53}{Simulated phenotype 53.}
  \item{PHENO54}{Simulated phenotype 54.}
  \item{PHENO55}{Simulated phenotype 55.}
  \item{PHENO56}{Simulated phenotype 56.}
  \item{PHENO57}{Simulated phenotype 57.}
  \item{PHENO58}{Simulated phenotype 58.}
  \item{PHENO59}{Simulated phenotype 59.}
  \item{PHENO60}{Simulated phenotype 60.}
  \item{PHENO61}{Simulated phenotype 61.}
  \item{PHENO62}{Simulated phenotype 62.}
  \item{PHENO63}{Simulated phenotype 63.}
  \item{PHENO64}{Simulated phenotype 64.}
  \item{PHENO65}{Simulated phenotype 65.}
  \item{PHENO66}{Simulated phenotype 66.}
  \item{PHENO67}{Simulated phenotype 67.}
  \item{PHENO68}{Simulated phenotype 68.}
  \item{PHENO69}{Simulated phenotype 69.}
  \item{PHENO70}{Simulated phenotype 70.}
  \item{PHENO71}{Simulated phenotype 71.}
  \item{PHENO72}{Simulated phenotype 72.}
  \item{PHENO73}{Simulated phenotype 73.}
  \item{PHENO74}{Simulated phenotype 74.}
  \item{PHENO75}{Simulated phenotype 75.}
  \item{PHENO76}{Simulated phenotype 76.}
  \item{PHENO77}{Simulated phenotype 77.}
  \item{PHENO78}{Simulated phenotype 78.}
  \item{PHENO79}{Simulated phenotype 79.}
  \item{PHENO80}{Simulated phenotype 80.}
  \item{PHENO81}{Simulated phenotype 81.}
  \item{PHENO82}{Simulated phenotype 82.}
  \item{PHENO83}{Simulated phenotype 83.}
  \item{PHENO84}{Simulated phenotype 84.}
  \item{PHENO85}{Simulated phenotype 85.}
  \item{PHENO86}{Simulated phenotype 86.}
  \item{PHENO87}{Simulated phenotype 87.}
  \item{PHENO88}{Simulated phenotype 88.}
  \item{PHENO89}{Simulated phenotype 89.}
  \item{PHENO90}{Simulated phenotype 90.}
  \item{PHENO91}{Simulated phenotype 91.}
  \item{PHENO92}{Simulated phenotype 92.}
  \item{PHENO93}{Simulated phenotype 93.}
  \item{PHENO94}{Simulated phenotype 94.}
  \item{PHENO95}{Simulated phenotype 95.}
  \item{PHENO96}{Simulated phenotype 96.}
  \item{PHENO97}{Simulated phenotype 97.}
  \item{PHENO98}{Simulated phenotype 98.}
  \item{PHENO99}{Simulated phenotype 99.}
  \item{PHENO100}{Simulated phenotype 100.}
  \item{PHENO101}{Simulated phenotype 101.}
  \item{PHENO102}{Simulated phenotype 102.}
  \item{PHENO103}{Simulated phenotype 103.}
  \item{PHENO104}{Simulated phenotype 104.}
  \item{PHENO105}{Simulated phenotype 105.}
  \item{PHENO106}{Simulated phenotype 106.}
  \item{PHENO107}{Simulated phenotype 107.}
  \item{PHENO108}{Simulated phenotype 108.}
  \item{PHENO109}{Simulated phenotype 109.}
  \item{PHENO110}{Simulated phenotype 110.}
  \item{PHENO111}{Simulated phenotype 111.}
  \item{PHENO112}{Simulated phenotype 112.}
  \item{PHENO113}{Simulated phenotype 113.}
  \item{PHENO114}{Simulated phenotype 114.}
  \item{PHENO115}{Simulated phenotype 115.}
  \item{PHENO116}{Simulated phenotype 116.}
  \item{PHENO117}{Simulated phenotype 117.}
  \item{PHENO118}{Simulated phenotype 118.}
  \item{PHENO119}{Simulated phenotype 119.}
  \item{PHENO120}{Simulated phenotype 120.}
  \item{PHENO121}{Simulated phenotype 121.}
  \item{PHENO122}{Simulated phenotype 122.}
  \item{PHENO123}{Simulated phenotype 123.}
  \item{PHENO124}{Simulated phenotype 124.}
  \item{PHENO125}{Simulated phenotype 125.}
  \item{PHENO126}{Simulated phenotype 126.}
  \item{PHENO127}{Simulated phenotype 127.}
  \item{PHENO128}{Simulated phenotype 128.}
  \item{PHENO129}{Simulated phenotype 129.}
  \item{PHENO130}{Simulated phenotype 130.}
  \item{PHENO131}{Simulated phenotype 131.}
  \item{PHENO132}{Simulated phenotype 132.}
  \item{PHENO133}{Simulated phenotype 133.}
  \item{PHENO134}{Simulated phenotype 134.}
  \item{PHENO135}{Simulated phenotype 135.}
  \item{PHENO136}{Simulated phenotype 136.}
  \item{PHENO137}{Simulated phenotype 137.}
  \item{PHENO138}{Simulated phenotype 138.}
  \item{PHENO139}{Simulated phenotype 139.}
  \item{PHENO140}{Simulated phenotype 140.}
  \item{PHENO141}{Simulated phenotype 141.}
  \item{PHENO142}{Simulated phenotype 142.}
  \item{PHENO143}{Simulated phenotype 143.}
  \item{PHENO144}{Simulated phenotype 144.}
  \item{PHENO145}{Simulated phenotype 145.}
  \item{PHENO146}{Simulated phenotype 146.}
  \item{PHENO147}{Simulated phenotype 147.}
  \item{PHENO148}{Simulated phenotype 148.}
  \item{PHENO149}{Simulated phenotype 149.}
  \item{PHENO150}{Simulated phenotype 150.}
  \item{PHENO151}{Simulated phenotype 151.}
  \item{PHENO152}{Simulated phenotype 152.}
  \item{PHENO153}{Simulated phenotype 153.}
  \item{PHENO154}{Simulated phenotype 154.}
  \item{PHENO155}{Simulated phenotype 155.}
  \item{PHENO156}{Simulated phenotype 156.}
  \item{PHENO157}{Simulated phenotype 157.}
  \item{PHENO158}{Simulated phenotype 158.}
  \item{PHENO159}{Simulated phenotype 159.}
  \item{PHENO160}{Simulated phenotype 160.}
  \item{PHENO161}{Simulated phenotype 161.}
  \item{PHENO162}{Simulated phenotype 162.}
  \item{PHENO163}{Simulated phenotype 163.}
  \item{PHENO164}{Simulated phenotype 164.}
  \item{PHENO165}{Simulated phenotype 165.}
  \item{PHENO166}{Simulated phenotype 166.}
  \item{PHENO167}{Simulated phenotype 167.}
  \item{PHENO168}{Simulated phenotype 168.}
  \item{PHENO169}{Simulated phenotype 169.}
  \item{PHENO170}{Simulated phenotype 170.}
  \item{PHENO171}{Simulated phenotype 171.}
  \item{PHENO172}{Simulated phenotype 172.}
  \item{PHENO173}{Simulated phenotype 173.}
  \item{PHENO174}{Simulated phenotype 174.}
  \item{PHENO175}{Simulated phenotype 175.}
  \item{PHENO176}{Simulated phenotype 176.}
  \item{PHENO177}{Simulated phenotype 177.}
  \item{PHENO178}{Simulated phenotype 178.}
  \item{PHENO179}{Simulated phenotype 179.}
  \item{PHENO180}{Simulated phenotype 180.}
  \item{PHENO181}{Simulated phenotype 181.}
  \item{PHENO182}{Simulated phenotype 182.}
  \item{PHENO183}{Simulated phenotype 183.}
  \item{PHENO184}{Simulated phenotype 184.}
  \item{PHENO185}{Simulated phenotype 185.}
  \item{PHENO186}{Simulated phenotype 186.}
  \item{PHENO187}{Simulated phenotype 187.}
  \item{PHENO188}{Simulated phenotype 188.}
  \item{PHENO189}{Simulated phenotype 189.}
  \item{PHENO190}{Simulated phenotype 190.}
  \item{PHENO191}{Simulated phenotype 191.}
  \item{PHENO192}{Simulated phenotype 192.}
  \item{PHENO193}{Simulated phenotype 193.}
  \item{PHENO194}{Simulated phenotype 194.}
  \item{PHENO195}{Simulated phenotype 195.}
  \item{PHENO196}{Simulated phenotype 196.}
  \item{PHENO197}{Simulated phenotype 197.}
  \item{PHENO198}{Simulated phenotype 198.}
  \item{PHENO199}{Simulated phenotype 199.}
  \item{PHENO200}{Simulated phenotype 200.}
  \item{PHENO201}{Simulated phenotype 201.}
  \item{PHENO202}{Simulated phenotype 202.}
  \item{PHENO203}{Simulated phenotype 203.}
  \item{PHENO204}{Simulated phenotype 204.}
  \item{PHENO205}{Simulated phenotype 205.}
  \item{PHENO206}{Simulated phenotype 206.}
  \item{PHENO207}{Simulated phenotype 207.}
  \item{PHENO208}{Simulated phenotype 208.}
  \item{PHENO209}{Simulated phenotype 209.}
  \item{PHENO210}{Simulated phenotype 210.}
  \item{PHENO211}{Simulated phenotype 211.}
  \item{PHENO212}{Simulated phenotype 212.}
  \item{PHENO213}{Simulated phenotype 213.}
  \item{PHENO214}{Simulated phenotype 214.}
  \item{PHENO215}{Simulated phenotype 215.}
  \item{PHENO216}{Simulated phenotype 216.}
  \item{PHENO217}{Simulated phenotype 217.}
  \item{PHENO218}{Simulated phenotype 218.}
  \item{PHENO219}{Simulated phenotype 219.}
  \item{PHENO220}{Simulated phenotype 220.}
  \item{PHENO221}{Simulated phenotype 221.}
  \item{PHENO222}{Simulated phenotype 222.}
  \item{PHENO223}{Simulated phenotype 223.}
  \item{PHENO224}{Simulated phenotype 224.}
  \item{PHENO225}{Simulated phenotype 225.}
  \item{PHENO226}{Simulated phenotype 226.}
  \item{PHENO227}{Simulated phenotype 227.}
  \item{PHENO228}{Simulated phenotype 228.}
  \item{PHENO229}{Simulated phenotype 229.}
  \item{PHENO230}{Simulated phenotype 230.}
  \item{PHENO231}{Simulated phenotype 231.}
  \item{PHENO232}{Simulated phenotype 232.}
  \item{PHENO233}{Simulated phenotype 233.}
  \item{PHENO234}{Simulated phenotype 234.}
  \item{PHENO235}{Simulated phenotype 235.}
  \item{PHENO236}{Simulated phenotype 236.}
  \item{PHENO237}{Simulated phenotype 237.}
  \item{PHENO238}{Simulated phenotype 238.}
  \item{PHENO239}{Simulated phenotype 239.}
  \item{PHENO240}{Simulated phenotype 240.}
  \item{PHENO241}{Simulated phenotype 241.}
  \item{PHENO242}{Simulated phenotype 242.}
  \item{PHENO243}{Simulated phenotype 243.}
  \item{PHENO244}{Simulated phenotype 244.}
  \item{PHENO245}{Simulated phenotype 245.}
  \item{PHENO246}{Simulated phenotype 246.}
  \item{PHENO247}{Simulated phenotype 247.}
  \item{PHENO248}{Simulated phenotype 248.}
  \item{PHENO249}{Simulated phenotype 249.}
  \item{PHENO250}{Simulated phenotype 250.}
  \item{PHENO251}{Simulated phenotype 251.}
  \item{PHENO252}{Simulated phenotype 252.}
  \item{PHENO253}{Simulated phenotype 253.}
  \item{PHENO254}{Simulated phenotype 254.}
  \item{PHENO255}{Simulated phenotype 255.}
  \item{PHENO256}{Simulated phenotype 256.}
  \item{PHENO257}{Simulated phenotype 257.}
  \item{PHENO258}{Simulated phenotype 258.}
  \item{PHENO259}{Simulated phenotype 259.}
  \item{PHENO260}{Simulated phenotype 260.}
  \item{PHENO261}{Simulated phenotype 261.}
  \item{PHENO262}{Simulated phenotype 262.}
  \item{PHENO263}{Simulated phenotype 263.}
  \item{PHENO264}{Simulated phenotype 264.}
  \item{PHENO265}{Simulated phenotype 265.}
  \item{PHENO266}{Simulated phenotype 266.}
  \item{PHENO267}{Simulated phenotype 267.}
  \item{PHENO268}{Simulated phenotype 268.}
  \item{PHENO269}{Simulated phenotype 269.}
  \item{PHENO270}{Simulated phenotype 270.}
  \item{PHENO271}{Simulated phenotype 271.}
  \item{PHENO272}{Simulated phenotype 272.}
  \item{PHENO273}{Simulated phenotype 273.}
  \item{PHENO274}{Simulated phenotype 274.}
  \item{PHENO275}{Simulated phenotype 275.}
  \item{PHENO276}{Simulated phenotype 276.}
  \item{PHENO277}{Simulated phenotype 277.}
  \item{PHENO278}{Simulated phenotype 278.}
  \item{PHENO279}{Simulated phenotype 279.}
  \item{PHENO280}{Simulated phenotype 280.}
  \item{PHENO281}{Simulated phenotype 281.}
  \item{PHENO282}{Simulated phenotype 282.}
  \item{PHENO283}{Simulated phenotype 283.}
  \item{PHENO284}{Simulated phenotype 284.}
  \item{PHENO285}{Simulated phenotype 285.}
  \item{PHENO286}{Simulated phenotype 286.}
  \item{PHENO287}{Simulated phenotype 287.}
  \item{PHENO288}{Simulated phenotype 288.}
  \item{PHENO289}{Simulated phenotype 289.}
  \item{PHENO290}{Simulated phenotype 290.}
  \item{PHENO291}{Simulated phenotype 291.}
  \item{PHENO292}{Simulated phenotype 292.}
  \item{PHENO293}{Simulated phenotype 293.}
  \item{PHENO294}{Simulated phenotype 294.}
  \item{PHENO295}{Simulated phenotype 295.}
  \item{PHENO296}{Simulated phenotype 296.}
  \item{PHENO297}{Simulated phenotype 297.}
  \item{PHENO298}{Simulated phenotype 298.}
  \item{PHENO299}{Simulated phenotype 299.}
  \item{PHENO300}{Simulated phenotype 300.}
}
}
\source{
<https://doi.org/10.21203/rs.3.rs-873449/v1>
}
\usage{
sipPair_exampleData
}
\description{
A sample-by-variable data frame with IDs,
covariates, and phenotypes as column names.
}
\keyword{datasets}
