% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sipPair_relatednessData}
\alias{sipPair_relatednessData}
\title{Single-iteration paired permutation relatedness data}
\format{
A data frame with 9886 rows and 4 variables:
\describe{
  \item{IID1}{Individual ID for one sample.}
  \item{IID2}{Iindividual ID for a second sample.}
  \item{PropIBD}{Identity-by-descent proportion of genome shared
  between the two samples.}
}
}
\source{
<https://doi.org/10.21203/rs.3.rs-873449/v1>
}
\usage{
sipPair_relatednessData
}
\description{
A sample-by-variable data frame with IDs
and identity-by-descent data between samples.
}
\keyword{datasets}
