% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sipPair_exampleData}
\alias{sipPair_exampleData}
\title{Single-iteration paired permutation phenotype and covariate data}
\format{
A data frame with 10,000 rows and 309 variables:
\describe{
  \item{FID}{Family ID.}
  \item{IID}{Individual ID.}
  \item{BATCH}{Genotyping batch.}
  \item{SEX}{Sample biological sex.}
  \item{AGE}{Sample age.}
  \item{STUDY}{Study contributing sample data.}
  \item{PCs}{4 genetic principal components.}
  \item{PHENOs}{300 simulated phenotypes.}
}
}
\source{
<https://doi.org/10.21203/rs.3.rs-873449/v1>
}
\usage{
sipPair_exampleData
}
\description{
A sample-by-variable data frame with IDs,
covariates, and phenotypes as column names.
}
\keyword{datasets}
